[Image support]
Implemented adding images when registering accommodation using the file picker.
When images are selected, list of all selected images is shown in the Datagrid.
When selecting an image from Datagrid, it's preview is shown in the menu.
After clicking Register button, selected images are copied to Resources/UserImages/ folder, and their file path is stored in ImagesURLs property of Accommodation class. Every Accommodation gets it's own folder 'Accommodation_ID'.
List of Accommodations is reduced, and images for all of them are prepared.