[Created Dependency Injector and an example of using DI with Repositories and Services]
Implemented a Dependency Injector in the class AppInjector (the name can be changed). Created repository interfaces for User and Accommodation, as well as their corresponding implementations.

The SignIn window has been adapted to use DI and the new repositories.

Additionally, a separate project has been created to provide an example of using DI with two different repositories.