[Feature Display Owner Ratings Implemented]
A button has been created in the OwnerHome window to display the guest rating for the selected reservation. The button will be disabled if no rating has been left or if the owner is not eligible to rate the guest. The button will be enabled if the owner's deadline for rating the guest has passed, or if the owner has already rated the guest.

The rest of the architecture for using the OwnerRating model has been created, including the Service, Repository, IRepo, and ViewModel. Several examples of guest ratings have also been created and linked to the reservation data.