[Super Owner Functionality Implemented]
The display of information regarding the owner's ratings and status has been implemented. A Popup window for displaying this information is accessed from the OwnerHomeView. Upon startup and login, the data is updated and recorded in a CSV file. Then, when initializing the Popup, the appropriate icons and information about the owner are displayed.

An IconsDictionary.xaml file has been added to define XAML-generated icons. An example of its usage can be found in the OwnerAccountPopupView.xaml file.