[Stats PDF Export implemented]
In every Stats view, new Share icon is added to top right corner. When pressed, you get prompted to select where to save file, and to give it a name. After that, depending on from which menu you pressed Share, PDF stats will be generated. It can be monthly or yearly, as stats in the app.